Bug fixed - Double Logs
Also changed how essential flags are called (now are executed rather than imported)
