+++ CODE/Logicytics.py
@@ -1,12 +1,11 @@
 import threading
-from _extra import menu, unzip
+
+from __lib_class import *
 from _health import backup, update
 from _hide_my_tracks import attempt_hide
 from _zipper import Zip
-from __lib_class import *
 
 log = Log(debug=DEBUG)
-
 
 """
 This python script is the main entry point for the tool called Logicytics.
@@ -44,29 +43,29 @@ else:
     input("Press Enter to exit...")
     exit(1)
 
+
+def special_run(file_path: str):
+    sr_current_dir = os.path.dirname(os.path.abspath(__file__))
+    sr_script_path = os.path.join(sr_current_dir, file_path)
+    sr_process = subprocess.Popen(["cmd.exe", "/c", "start", "python", sr_script_path])
+    sr_process.wait()
+    exit(0)
+
+
 # Special actions -> Quit
 if action == "debug":
-    current_dir = os.path.dirname(os.path.abspath(__file__))
-    script_path = os.path.join(current_dir, "_debug.py")
-    process = subprocess.Popen(["cmd.exe", "/c", "start", "python", script_path])
-    process.wait()
-    input("Press Enter to exit...")
-    exit(0)
+    log.info("Opening debug menu...")
+    special_run("_debug.py")
 
 check_status.sys_internal_zip()
 
 if action == "dev":
-    current_dir = os.path.dirname(os.path.abspath(__file__))
-    script_path = os.path.join(current_dir, "_dev.py")
-    process = subprocess.Popen(["cmd.exe", "/c", "start", "python", script_path])
-    process.wait()
-    exit(0)
+    log.info("Opening developer menu...")
+    special_run("_dev.py")
 
 if action == "extra":
     log.info("Opening extra tools menu...")
-    menu()
-    input("Press Enter to exit...")
-    exit(0)
+    special_run("_extra.py")
 
 if action == "update":
     log.info("Updating...")
@@ -100,14 +99,12 @@ if action == "unzip_extra":
         "caution"
     )
     log.info("Unzipping...")
-    unzip("..\\EXTRA\\EXTRA.zip")
+    Actions().unzip("..\\EXTRA\\EXTRA.zip")
     log.info("Unzip complete!")
     input("Press Enter to exit...")
     exit(0)
 
-
 log.info("Starting Logicytics...")
-
 
 # Check for privileges and errors
 if not check_status.admin():
@@ -163,7 +160,6 @@ if action == "exe":
 if action == "modded":
     # Add all files in MODS to execution list
     execution_list = Execute.get_files("../MODS", execution_list)
-
 
 log.debug(execution_list)
